Delete unused code; Restructure components; Implement add tracks functionality
