+++ rollup.config.js
@@ -67,6 +67,7 @@ export default {
     typescript({
       sourceMap: !production,
       inlineSources: !production,
+      rootDir: "./src", // fixes import type issue
     }),
 
     // In dev mode, call `npm run start` once
