--- src/LyricsTrack.svelte
@@ -1,10 +1,0 @@
-<script>
-import Track from "./Track.svelte";
-
-</script>
-<div>
-  <h2>Lyrics</h2>
-  <Track>
-    <slot />
-  </Track>
-</div>
