+++ src/constants/midi.ts
@@ -1,2 +1,1 @@
-export const BPM_DIVISOR = 60000000;
 export const WHOLE_NOTE_TICKS = 384;
