--- src/midi/keySignature.ts
@@ -1,9 +1,0 @@
-// Key signature
-// FF 59 02 sf mi
-// The key signature is specified by the numeric sf Key value,
-/// which is 0 for the key of C, a positive value for each sharp
-// above C, or a negative value for each flat below C, thus in
-// the inclusive range -7 to 7. The Major/Minor mi field is a number
-// value which will be 0 for a major key and 1 for a minor key.
-
-export function keySignatureToMidi() {}
