--- src/newStructure.js
@@ -1,32 +1,0 @@
-const structure = {
-  song: "Teenage",
-  tracks: [
-    {
-      type: "lyrics",
-      originalLyric: "Her name is Noelle",
-      nodes: [
-        { data: "Her name is Noelle", position: 0, duration: 384 },
-        { data: "I had a dream about her", position: 0, duration: 384 },
-      ],
-    },
-    {
-      type: "chords",
-      nodes: [
-        { data: "E", position: 48, duration: 96 },
-        { data: "B", position: 144, duration: 96 },
-        { data: "E", position: 48, duration: 96 },
-        { data: "A", position: 144, duration: 96 },
-      ],
-    },
-    {
-      type: "lead-vocal",
-      nodes: [
-        { data: "G#4", position: 48, duration: 48 },
-        { data: "G#4", position: 48, duration: 48 },
-        { data: "G#4", position: 48, duration: 48 },
-        { data: "G#4", position: 48, duration: 48 },
-        { data: "G#4", position: 48, duration: 48 },
-      ],
-    },
-  ],
-};
